Remove extra container ApplciationNew to simplify creation
