+++ src/containers/Application/index.ts
@@ -1,5 +1,4 @@
 import ApplicationCreate from './ApplicationCreate'
-import ApplicationNew from './ApplicationNew'
 import ApplicationPage from './ApplicationPage'
 
-export { ApplicationCreate, ApplicationNew, ApplicationPage }
+export { ApplicationCreate, ApplicationPage }
